Fusión Silvio-Martín nivel 3

+++ AD2015C2_Server/src/dominio/ItemSC.java
@@ -6,10 +6,10 @@ import entities.ItemSCENT;
 public class ItemSC {
 	
 	private Rodamiento rodamiento;
-	private float cantidad;
+	private int cantidad;
 	private float precio;
 	
-	public ItemSC(Rodamiento rodamiento, float cantidad, float precio) {
+	public ItemSC(Rodamiento rodamiento, int cantidad, float precio) {
 		super();
 		this.rodamiento = rodamiento;
 		this.cantidad = cantidad;
@@ -25,11 +25,11 @@ public class ItemSC {
 		this.rodamiento = rodamiento;
 	}
 
-	public float getCantidad() {
+	public int getCantidad() {
 		return cantidad;
 	}
 
-	public void setCantidad(float cantidad) {
+	public void setCantidad(int cantidad) {
 		this.cantidad = cantidad;
 	}
